# doc/api.yaml
openapi: 3.1.0
info:
  title: Go Clean Starter API
  description: This is an Open API 3.1.0 specification document for Go Clean Starter API
  version: "1.0.0"

servers:
  - description: Local
    url: http://localhost:8080

tags:
  - name: health
    description: Health check endpoints
  - name: users
    description: User management operations

paths:
  /health:
    get:
      summary: Health check
      description: Check if the API is running and healthy
      tags:
        - health
      operationId: getHealthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "OK"

  /users:
    post:
      summary: Create a new user
      description:  Create a new user with name, email, and password
      tags:
        - users
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          $ref: '#/components/responses/400'
        '409':
          $ref: '#/components/responses/409'
        '500':
          $ref: '#/components/responses/500'

  /users/{id}:
    get:
      summary: Get user by ID
      description: This endpoint returns a user by their UUID
      tags:
        - users
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/user_id'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

components:

  responses:
    '400':
      description: 'Bad Request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    '401':
      description: 'Unauthorized'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    '404':
      description: 'Not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    '409':
      description: 'Conflict'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    '500':
      description: 'Internal Server Error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'

  parameters:
    user_id:
      name: id
      in: path
      required: true
      description: User UUID
      schema:
        type: string
        format: uuid
        example: "123e4567-e89b-12d3-a456-426614174000"


  schemas:
  ############################################################
  #                     REQUEST schemas
  ############################################################
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
          description: User's full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User's email address
          example: "john.doe@example.com"
        password:
          type: string
          description: User's password
          example: "securepassword123"
      required:
        - name
        - email
        - password

    ############################################################
    #                     RESPONSE schemas
    ############################################################
    CreateUserResponse:
      type: object
      description: User creation response
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the newly created user
          example: "123e4567-e89b-12d3-a456-426614174000"

    UserResponse:
      type: object
      description: User representation
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: User's full name
          example: "John Doe"

    ErrorMessage:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: 'Some error message'