# internal/http/handler/openapi_types.gen.go
// Package handler provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package handler

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// CreateUserRequest defines model for CreateUserRequest.
type CreateUserRequest struct {
	// Email User's email address
	Email openapi_types.Email `json:"email"`

	// Name User's full name
	Name string `json:"name"`

	// Password User's password
	Password string `json:"password"`
}

// CreateUserResponse User creation response
type CreateUserResponse struct {
	// Id Unique identifier for the newly created user
	Id openapi_types.UUID `json:"id"`
}

// ErrorMessage defines model for ErrorMessage.
type ErrorMessage struct {
	Message string `json:"message"`
}

// UserResponse User representation
type UserResponse struct {
	// Id Unique identifier for the user
	Id openapi_types.UUID `json:"id"`

	// Name User's full name
	Name string `json:"name"`
}

// UserId defines model for user_id.
type UserId = openapi_types.UUID

// N400 defines model for 400.
type N400 = ErrorMessage

// N404 defines model for 404.
type N404 = ErrorMessage

// N409 defines model for 409.
type N409 = ErrorMessage

// N500 defines model for 500.
type N500 = ErrorMessage

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody = CreateUserRequest